+++ lastversion/HolderFactory.py
@@ -50,7 +50,7 @@ class HolderFactory:
     @staticmethod
     # go through subclasses in order to find the one that is holding a given project
     # repo is either complete URL or a name allowing to identify a single project
-    def get_instance_for_repo(repo, only=None, at=None):
+    def get_instance_for_repo(repo, at=None):
         """Find the right hosting for this repo."""
         if at == 'helm_chart' or (at and '/' not in repo):
             return HolderFactory.HOLDERS[at](repo, hostname=None)
@@ -84,12 +84,14 @@ class HolderFactory:
                 raise BadProjectError(
                     'No project found. Could not guess a repo from homepage'
                 )
+            
         if known_repo and 'branches' in known_repo:
             holder.set_branches(known_repo['branches'])
+            
         if known_repo and 'only' in known_repo:
             holder.set_only(known_repo['only'])
-        if only:
-            holder.set_only(only)
+            
         if known_repo and 'release_url_format' in known_repo:
             holder.RELEASE_URL_FORMAT = known_repo['release_url_format']
+            
         return holder