+++ lastversion/spdx_id_to_rpmspec.py
@@ -1,0 +1,8 @@
+# Maps SPDX ID from https://spdx.org/licenses/
+# to corresponding short name license as recommended by Fedora
+# https://fedoraproject.org/wiki/Licensing:Main?rd=Licensing#Good_Licenses
+rpmspec_licenses = {
+   'Apache-2.0': 'APSL 2.0',
+   'BSD-2-Clause': 'BSD',
+   'MIT': 'MIT',
+}